add sext, ptrtoint and inttoptr

+++ include/passes/gvn.hpp
@@ -302,7 +302,7 @@ namespace GVNExpression
                 return false;
             if (operands.size() != other->operands.size())
                 return false;
-            for (int i = 0; i < operands.size(); i++)
+            for (size_t i = 0; i < operands.size(); i++)
             {
                 if (!(operands[i] == other->operands[i]))
                     return false;
@@ -341,7 +341,7 @@ namespace GVNExpression
                 return false;
             if (operands.size() != other->operands.size())
                 return false;
-            for (int i = 0; i < operands.size(); i++)
+            for (size_t i = 0; i < operands.size(); i++)
             {
                 if (!(operands[i] == other->operands[i]))
                     return false;
